chg: create feeds command shell, move out feed commands from servers shell
